# web/swagger/dist/doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a FFmpeg Go server.",
    "title": "FFmpeg Go API",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "version": "1.0"
  },
  "host": "localhost:8084",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
    "/monitor/stats": {
      "get": {
        "description": "获取系统资源使用情况统计信息，包括CPU、内存、磁盘等",
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "获取系统统计信息",
        "operationId": "get-system-stats",
        "responses": {
          "200": {
            "description": "系统统计信息",
            "schema": {
              "$ref": "#/definitions/SystemStats"
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks": {
      "get": {
        "description": "获取所有任务列表，支持按状态和优先级筛选",
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "获取任务列表",
        "operationId": "get-tasks",
        "parameters": [
          {
            "type": "string",
            "description": "任务状态筛选",
            "name": "status",
            "in": "query"
          },
          {
            "type": "string",
            "description": "任务优先级筛选",
            "name": "priority",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "任务列表",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Task"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks/stats": {
      "get": {
        "description": "获取任务统计信息，包括各种状态的任务数量",
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "获取任务统计信息",
        "operationId": "get-task-stats",
        "responses": {
          "200": {
            "description": "任务统计信息",
            "schema": {
              "$ref": "#/definitions/TaskStats"
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks/{taskId}": {
      "get": {
        "description": "根据任务ID获取任务详细信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "获取任务详情",
        "operationId": "get-task-detail",
        "parameters": [
          {
            "type": "string",
            "description": "任务ID",
            "name": "taskId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "任务详情",
            "schema": {
              "$ref": "#/definitions/Task"
            }
          },
          "404": {
            "description": "任务未找到",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks/{taskId}/executions": {
      "get": {
        "description": "获取指定任务的所有执行历史记录",
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "获取任务执行历史",
        "operationId": "get-task-executions",
        "parameters": [
          {
            "type": "string",
            "description": "任务ID",
            "name": "taskId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "任务执行历史记录列表",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/TaskExecution"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks/{taskId}/log": {
      "get": {
        "description": "获取指定任务的日志内容",
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "获取任务日志",
        "operationId": "get-task-log",
        "parameters": [
          {
            "type": "string",
            "description": "任务ID",
            "name": "taskId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "任务日志内容",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "任务日志未找到",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks/cancel": {
      "post": {
        "description": "取消一个待处理或处理中的任务",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "取消任务",
        "operationId": "cancel-task",
        "parameters": [
          {
            "description": "任务取消请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TaskCancelRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取消成功",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "请求参数错误或任务状态不正确",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "任务未找到",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks/discard": {
      "post": {
        "description": "丢弃一个已完成或失败的任务",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "丢弃任务",
        "operationId": "discard-task",
        "parameters": [
          {
            "description": "任务丢弃请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TaskDiscardRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "丢弃成功",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "请求参数错误或任务状态不正确",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "任务未找到",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/tasks/retry": {
      "post": {
        "description": "重试一个失败的任务",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "重试失败的任务",
        "operationId": "retry-task",
        "parameters": [
          {
            "description": "任务重试请求",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TaskRetryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "重试成功",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "404": {
            "description": "任务未找到",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/monitor/workers": {
      "get": {
        "description": "获取Worker池的统计信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "monitor"
        ],
        "summary": "获取Worker统计信息",
        "operationId": "get-worker-stats",
        "responses": {
          "200": {
            "description": "Worker统计信息",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "500": {
            "description": "内部服务器错误",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SystemStats": {
      "type": "object",
      "properties": {
        "activeWorkers": {
          "description": "活跃工作线程数",
          "type": "integer"
        },
        "cpuUsage": {
          "description": "CPU使用率",
          "type": "number"
        },
        "diskTotal": {
          "description": "总磁盘空间",
          "type": "integer"
        },
        "diskUsage": {
          "description": "磁盘使用率",
          "type": "number"
        },
        "diskUsed": {
          "description": "已使用磁盘空间",
          "type": "integer"
        },
        "goroutines": {
          "description": "Goroutines数量",
          "type": "integer"
        },
        "memoryTotal": {
          "description": "总内存",
          "type": "integer"
        },
        "memoryUsage": {
          "description": "内存使用率",
          "type": "number"
        },
        "memoryUsed": {
          "description": "已使用内存",
          "type": "integer"
        },
        "taskQueueSize": {
          "description": "任务队列大小",
          "type": "integer"
        },
        "timestamp": {
          "description": "时间戳",
          "type": "string"
        },
        "workerCount": {
          "description": "工作线程总数",
          "type": "integer"
        }
      }
    },
    "Task": {
      "type": "object",
      "properties": {
        "created": {
          "type": "string"
        },
        "error": {
          "type": "string"
        },
        "executionCount": {
          "type": "integer"
        },
        "finished": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "lastExecution": {
          "type": "string"
        },
        "priority": {
          "$ref": "#/definitions/TaskPriority"
        },
        "progress": {
          "type": "number"
        },
        "result": {
          "type": "string"
        },
        "spec": {
          "type": "object"
        },
        "started": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "TaskCancelRequest": {
      "type": "object",
      "properties": {
        "taskId": {
          "description": "任务ID",
          "type": "string"
        }
      },
      "required": [
        "taskId"
      ]
    },
    "TaskDiscardRequest": {
      "type": "object",
      "properties": {
        "taskId": {
          "description": "任务ID",
          "type": "string"
        }
      },
      "required": [
        "taskId"
      ]
    },
    "TaskExecution": {
      "type": "object",
      "properties": {
        "created": {
          "type": "string"
        },
        "error": {
          "type": "string"
        },
        "executionNumber": {
          "type": "integer"
        },
        "executionTime": {
          "type": "integer"
        },
        "finished": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "priority": {
          "$ref": "#/definitions/TaskPriority"
        },
        "progress": {
          "type": "number"
        },
        "result": {
          "type": "string"
        },
        "spec": {
          "type": "object"
        },
        "started": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "taskId": {
          "type": "string"
        }
      }
    },
    "TaskPriority": {
      "type": "integer",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "x-enum-varnames": [
        "PriorityLow",
        "PriorityNormal",
        "PriorityHigh",
        "PriorityCritical"
      ]
    },
    "TaskRetryRequest": {
      "type": "object",
      "properties": {
        "taskId": {
          "description": "任务ID",
          "type": "string"
        }
      },
      "required": [
        "taskId"
      ]
    },
    "TaskStats": {
      "type": "object",
      "properties": {
        "completedTasks": {
          "description": "已完成任务数",
          "type": "integer"
        },
        "failedTasks": {
          "description": "失败任务数",
          "type": "integer"
        },
        "pendingTasks": {
          "description": "待处理任务数",
          "type": "integer"
        },
        "processingTasks": {
          "description": "处理中任务数",
          "type": "integer"
        },
        "totalTasks": {
          "description": "总任务数",
          "type": "integer"
        }
      }
    },
...
  }
}